creating a python package
adding missing modules

changing imports

removing init

changing imports

installation as a package

creating a python package
