+++ examples/multi_gcnn_graph_classification_with_edge_features_example.py
@@ -1,6 +1,6 @@
-from keras.models import Input, Model, Sequential
-from keras.layers import Dense, Activation, Dropout, Lambda
-import keras.backend as K
+from tensorflow.keras.models import Input, Model, Sequential
+from tensorflow.keras.layers import Dense, Activation, Dropout, Lambda
+import tensorflow.keras.backend as K
 import numpy as np
 import pandas as pd
 from sklearn.utils import shuffle
