+++ keras_dgl/layers/__init__.py
@@ -1,5 +1,6 @@
+
 from .graph_cnn_layer import GraphCNN
 from .multi_graph_cnn_layer import MultiGraphCNN
 from .graph_attention_cnn_layer import GraphAttentionCNN
 from .multi_graph_attention_cnn_layer import MultiGraphAttentionCNN
-from .graph_convolutional_recurrent_layer import GraphConvLSTM
+from .graph_ops import graph_conv_op
