+++ keras_dgl/layers/graph_ops.py
@@ -1,4 +1,4 @@
-import keras.backend as K
+import tensorflow.keras.backend as K
 import tensorflow as tf
