+++ keras_dgl/layers/multi_graph_cnn_layer.py
@@ -1,7 +1,7 @@
-from keras import activations, initializers, constraints
-from keras import regularizers
-import keras.backend as K
-from keras.engine.topology import Layer
+from tensorflow.keras import activations, initializers, constraints
+from tensorflow.keras import regularizers
+import tensorflow.keras.backend as K
+from tensorflow.keras.layers import Layer
 import tensorflow as tf
 from .graph_ops import graph_conv_op
