+++ keras_dgl/utils/__init__.py
@@ -1,0 +1,1 @@
+from .loading_data import encode_onehot, load_data, load_data_attention
