+++ keras_dgl/utils/utils.py
@@ -2,70 +2,6 @@ from __future__ import print_function
 
 import scipy.sparse as sp
 import numpy as np
-from scipy.sparse.linalg.eigen.arpack import eigsh, ArpackNoConvergence
-
-
-def encode_onehot(labels):
-    classes = set(labels)
-    classes_dict = {c: np.identity(len(classes))[i, :] for i, c in enumerate(classes)}
-    labels_onehot = np.array(list(map(classes_dict.get, labels)), dtype=np.int32)
-    return labels_onehot
-
-
-def load_data(path="data/cora/", dataset="cora"):
-    """Load citation network dataset (cora only for now)"""
-    print('Loading {} dataset...'.format(dataset))
-
-    idx_features_labels = np.genfromtxt("{}{}.content".format(path, dataset), dtype=np.dtype(str))
-    features = sp.csr_matrix(idx_features_labels[:, 1:-1], dtype=np.float32)
-    labels = encode_onehot(idx_features_labels[:, -1])
-
-    # build graph
-    idx = np.array(idx_features_labels[:, 0], dtype=np.int32)
-    idx_map = {j: i for i, j in enumerate(idx)}
-    edges_unordered = np.genfromtxt("{}{}.cites".format(path, dataset), dtype=np.int32)
-    edges = np.array(list(map(idx_map.get, edges_unordered.flatten())),
-                     dtype=np.int32).reshape(edges_unordered.shape)
-    adj = sp.coo_matrix((np.ones(edges.shape[0]), (edges[:, 0], edges[:, 1])),
-                        shape=(labels.shape[0], labels.shape[0]), dtype=np.float32)
-
-    # build symmetric adjacency matrix
-    adj = adj + adj.T.multiply(adj.T > adj) - adj.multiply(adj.T > adj)
-
-    # features = normalize_features(features)
-    # adj = normalize_adj(adj + sp.eye(adj.shape[0]))
-
-    print('Dataset has {} nodes, {} edges, {} features.'.format(adj.shape[0], edges.shape[0], features.shape[1]))
-
-    return features.todense(), adj, labels
-
-
-def load_data_attention(path="data/cora/", dataset="cora"):
-    """Load citation network dataset (cora only for now)"""
-    print('Loading {} dataset...'.format(dataset))
-
-    idx_features_labels = np.genfromtxt("{}{}.content".format(path, dataset), dtype=np.dtype(str))
-    features = sp.csr_matrix(idx_features_labels[:, 1:-1], dtype=np.float32)
-    labels = encode_onehot(idx_features_labels[:, -1])
-
-    # build graph
-    idx = np.array(idx_features_labels[:, 0], dtype=np.int32)
-    idx_map = {j: i for i, j in enumerate(idx)}
-    edges_unordered = np.genfromtxt("{}{}.cites".format(path, dataset), dtype=np.int32)
-    edges = np.array(list(map(idx_map.get, edges_unordered.flatten())),
-                     dtype=np.int32).reshape(edges_unordered.shape)
-    adj = sp.coo_matrix((np.ones(edges.shape[0]), (edges[:, 0], edges[:, 1])),
-                        shape=(labels.shape[0], labels.shape[0]), dtype=np.float32)
-
-    # build symmetric adjacency matrix
-    adj = adj + adj.T.multiply(adj.T > adj) - adj.multiply(adj.T > adj)
-
-    features = normalize_features(features)
-    adj = normalize_adj(adj + sp.eye(adj.shape[0]))
-
-    print('Dataset has {} nodes, {} edges, {} features.'.format(adj.shape[0], edges.shape[0], features.shape[1]))
-
-    return features.todense(), adj, labels
 
 
 def normalize_features(mx):
@@ -226,18 +162,6 @@ def normalized_laplacian(adj, symmetric=True):
     adj_normalized = normalize_adj(adj, symmetric)
     laplacian = sp.eye(adj.shape[0]) - adj_normalized
     return laplacian
-
-
-def rescale_laplacian(laplacian):
-    try:
-        print('Calculating largest eigenvalue of normalized graph Laplacian...')
-        largest_eigval = eigsh(laplacian, 1, which='LM', return_eigenvectors=False)[0]
-    except ArpackNoConvergence:
-        print('Eigenvalue calculation did not converge! Using largest_eigval=2 instead.')
-        largest_eigval = 2
-
-    scaled_laplacian = (2. / largest_eigval) * laplacian - sp.eye(laplacian.shape[0])
-    return scaled_laplacian
 
 
 def chebyshev_polynomial(X, k):
